persona juridica test

+++ persona-model/persona-model-api/src/main/java/org/sistcoop/models/PersonaJuridicaProvider.java
@@ -28,7 +28,7 @@ public interface PersonaJuridicaProvider extends Provider {
 
 	List<PersonaJuridicaModel> getPersonasJuridicas();
 
-	int getPersonasJuridicasCount();
+	long getPersonasJuridicasCount();
 
 	List<PersonaJuridicaModel> getPersonasJuridicas(int firstResult,int maxResults);
